requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-generativeai>=0.3.0
python-dotenv
# hakushin: 사이트 폐쇄로 비활성화됨 (2026.02)
